Epochs store the time scale used at initialization
This is non-breaking change (surprisingly). It removes the
ceil_with_timesystem function created in the previous
unreleased 3.5.0-alpha (and same for floor and round) and now it always
performs these operations in the appropriate time scale.

This also adds durations to Epochs in the appropriate time scale.

TimeSystem has been renamed to TimeScale to match the literature (only
ESA NaviPedia uses the term `TimeSystem`).

Finally, RFC3339 is also supported when initializing from a gregorian
string.