Non-leap time scales error on init with leap second
Previously, one could initialize an Epoch with the 60th second on a leap
day even if the selected time scale did not support leap seconds.
This PR makes this an error.

Importantly, this PR also fixes an error in the `julian_epoch` test
where an epoch was initialized on the leap second but ... in TAI.
There is no leap second in TAI, so this initialization should fail.
The test case was also expanded to confirm the computation of Modified
Julian Dates using NASA HEASARC, with an extra test and a clarification
that leap second management is implementation dependent.