feat: replace `std::time::Duration` with `core::time::Duration`
the existing type conversions are between `hifitime::Duration` and
`std::time::Duration`, but don't support converting to the new (as of
Rust 1.25.0)
[`core::time::Duration`](https://doc.rust-lang.org/stable/core/time/struct.Duration.html).
this means that `no_std` usage of `hifitime` is missing some ergonomics.

to address this, replace the existing implementation and swap usage of
`std::time::Duration` with `core::time::Duration`, which should be
functionally equivalent, and remove the conditional compilation (gated
behind the `std` feature flag).

the MSRV of `hifitime` should be _at least_ 1.59.0, with v4.0.0
requiring 1.77.0, both of which should contain `core::time::Duration`.

fixes #439